requirements.txt
streamlit
youtube-transcript-api
google-generativeai
markdown-pdf
streamlit-lottie
requests